LayoutTests/imported/w3c/web-platform-tests/trusted-types/GlobalEventHandlers-onclick-expected.txt

CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"

PASS a.setAttribte('onclick') sets a trusted script.
PASS a.setAttribute('onclick') sets an unsuitable trusted type.
PASS a.setAttribute('click') sets a test string.
